Moved Functions and created function base class
